feat: add support for choosing OpenAI model
Add a new `Model` field to the `CLI` struct to allow users to choose the
OpenAI model to use. The default value is `text-davinci-003`.

In the `openai.go` file, the `createCompletion` function now uses the
model specified in the `Service` struct. The `Temperature` and `TopP`
parameters have been adjusted to new values. A new comment has been
added to remind the developer to find optimal values for these
parameters. The prompt message has been updated to include more
specific instructions on how to write the documentation.