feat(finder): add support for finding all code
Add a new method `All()` to the `Finder` struct that returns all the findings in the repository. This method is called by the `find()` method with the `all` parameter set to `true`.

Also, modify the `Uncommented()` method to call the `find()` method with the `all` parameter set to `false`.

Finally, rename the `findUncommented()` method to `findInFile()` to reflect its new functionality.