chore(jotbot-ts): refactor release script to use async/await and add support for git tagging and pushing
This commit refactors the release script to use async/await instead of execSync and adds support for git tagging and pushing. The script now uses the promisify function from the util module to convert the exec function to a promise-based function. The execute function is now an async function that uses the promisified exec function to execute commands. The script now adds support for git tagging and pushing by adding a new commit with a message that includes the new version number and pushing the changes to the remote repository.